Fix malformed BioProject ID regex
Previously, the regex used to separate projects into single and paired
end assumed that all BioProjects were of the form PRJNA. This excludes
BioProjects from other databases (e.g. PRJEB, PRJDB) and was resulting
in issues with the downstream results.

Bad regex results in a tuple of the format ("???", []) which messed up
the assembly step with null paths.

h/t @yttria-aniseia